feature: update doris instance detail web
