+++ scaleph-ui-react/src/pages/Project/Workspace/Doris/OperatorInstance/Detail/index.tsx
@@ -1,224 +1,27 @@
 import {useIntl, useLocation} from "umi";
-import React, {useEffect, useRef, useState} from "react";
+import React from "react";
 import {WsDorisOperatorInstance} from "@/services/project/typings";
-import {PageContainer, ProCard, StatisticCard} from "@ant-design/pro-components";
-import {Button, Divider, message, Popconfirm, Space, Statistic} from "antd";
-import RcResizeObserver from 'rc-resize-observer';
-import Editor, {Monaco, useMonaco} from "@monaco-editor/react";
-import YAML from "yaml";
-import {WsDorisOperatorInstanceService} from "@/services/project/WsDorisOperatorInstanceService";
-import {CaretRightOutlined, CloseOutlined} from "@ant-design/icons";
+import {PageContainer} from "@ant-design/pro-components";
+import {Divider} from "antd";
+import DorisInstanceDetailComponent
+  from "@/pages/Project/Workspace/Doris/OperatorInstance/Detail/DorisInstanceComponent";
+import DorisInstanceDetailYAML from "@/pages/Project/Workspace/Doris/OperatorInstance/Detail/DorisInstanceYaml";
+import DorisInstanceDetailAction from "@/pages/Project/Workspace/Doris/OperatorInstance/Detail/DorisInstanceAction";
+import DorisInstanceDetailAccess from "@/pages/Project/Workspace/Doris/OperatorInstance/Detail/DorisInstanceAccess";
 
 const DorisInstanceDetailWeb: React.FC = () => {
   const intl = useIntl();
   const data = useLocation().state as WsDorisOperatorInstance
-  const editorRef = useRef(null);
-  const monaco = useMonaco();
-
-  const [yaml, setYaml] = useState<string>(null);
-  const [responsive, setResponsive] = useState(false);
-
-  useEffect(() => {
-    monaco?.languages.typescript.javascriptDefaults.setEagerModelSync(true);
-  }, [monaco]);
-
-  const handleEditorDidMount = (editor, monaco: Monaco) => {
-    editorRef.current = editor;
-  }
-
-  useEffect(() => {
-    const yamlData = {...data}
-    yamlData.deployed = undefined
-    WsDorisOperatorInstanceService.asYaml(yamlData).then((response) => {
-      if (response.success) {
-        setYaml(YAML.stringify(response.data))
-      }
-    });
-  }, []);
 
   return (
     <PageContainer title={intl.formatMessage({id: 'pages.project.doris.instance.detail'})}>
-      <RcResizeObserver
-        key="resize-observer"
-        onResize={(offset) => {
-          setResponsive(offset.width < 596);
-        }}
-      >
-        <ProCard.Group direction={responsive ? 'column' : 'row'}>
-          <ProCard extra={
-            <div>
-              <Popconfirm
-                title={intl.formatMessage({id: 'app.common.operate.submit.confirm.title'})}
-                onConfirm={() => {
-                  WsDorisOperatorInstanceService.deploy(data.id).then(response => {
-                    if (response.success) {
-                      message.success(intl.formatMessage({id: 'app.common.operate.submit.success'}));
-                    }
-                  })
-                }}
-              >
-                <Button type="default" icon={<CaretRightOutlined/>}>
-                  {intl.formatMessage({id: 'pages.project.doris.instance.detail.deploy'})}
-                </Button>
-              </Popconfirm>
-
-
-              <Popconfirm
-                title={intl.formatMessage({id: 'app.common.operate.submit.confirm.title'})}
-                onConfirm={() => {
-                  WsDorisOperatorInstanceService.shutdown(data.id).then(response => {
-                    if (response.success) {
-                      message.success(intl.formatMessage({id: 'app.common.operate.submit.success'}));
-                    }
-                  })
-                }}
-              >
-                <Button icon={<CloseOutlined/>}>
-                  {intl.formatMessage({id: 'pages.project.doris.instance.detail.shutdown'})}
-                </Button>
-              </Popconfirm>
-            </div>
-          }>
-            集群信息
-          </ProCard>
-        </ProCard.Group>
-        <Divider type={'horizontal'}/>
-        <ProCard.Group title={intl.formatMessage({id: 'pages.project.doris.instance.detail.component'})}
-                       direction={responsive ? 'column' : 'row'}>
-          <ProCard bordered hoverable>
-            <StatisticCard title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.fe'})}
-                           statistic={{
-                             value: data.feSpec?.replicas,
-                             prefix: intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.replicas'}) + ' x',
-                             description: (
-                               <Space>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.requests.cpu'})}
-                                   value={data.feSpec?.requests?.cpu}/>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.requests.memory'})}
-                                   value={data.feSpec?.requests?.memory}/>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.limits.cpu'})}
-                                   value={data.feSpec?.limits?.cpu}/>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.limits.memory'})}
-                                   value={data.feSpec?.limits?.memory}/>
-                               </Space>
-                             ),
-                           }}
-                           footer={
-                             <div>{intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.image'}) + ": " + data.feSpec?.image ? data.feSpec?.image : '-'}</div>
-                           }
-            />
-          </ProCard>
-          <Divider type={responsive ? 'horizontal' : 'vertical'}/>
-          <ProCard bordered hoverable>
-            <StatisticCard title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.be'})}
-                           statistic={{
-                             value: data.beSpec?.replicas,
-                             prefix: intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.replicas'}) + ' x',
-                             description: (
-                               <Space>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.requests.cpu'})}
-                                   value={data.beSpec?.requests?.cpu}/>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.requests.memory'})}
-                                   value={data.beSpec?.requests?.memory}/>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.limits.cpu'})}
-                                   value={data.beSpec?.limits?.cpu}/>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.limits.memory'})}
-                                   value={data.beSpec?.limits?.memory}/>
-                               </Space>
-                             ),
-                           }}
-                           footer={
-                             <div>{intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.image'}) + ": " + data.beSpec?.image ? data.beSpec?.image : '-'}</div>
-                           }
-            />
-          </ProCard>
-          <Divider type={responsive ? 'horizontal' : 'vertical'}/>
-          <ProCard bordered hoverable>
-            <StatisticCard title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.cn'})}
-                           statistic={{
-                             value: data.cnSpec?.replicas,
-                             prefix: intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.replicas'}) + ' x',
-                             description: (
-                               <Space>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.requests.cpu'})}
-                                   value={data.cnSpec?.requests?.cpu}/>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.requests.memory'})}
-                                   value={data.cnSpec?.requests?.memory}/>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.limits.cpu'})}
-                                   value={data.cnSpec?.limits?.cpu}/>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.limits.memory'})}
-                                   value={data.cnSpec?.limits?.memory}/>
-                               </Space>
-                             ),
-                           }}
-                           footer={
-                             <div>{intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.image'}) + ": " + data.cnSpec?.image ? data.cnSpec?.image : '-'}</div>
-                           }
-            />
-          </ProCard>
-          <Divider type={responsive ? 'horizontal' : 'vertical'}/>
-          <ProCard bordered hoverable>
-            <StatisticCard title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.broker'})}
-                           statistic={{
-                             value: data.brokerSpec?.replicas,
-                             prefix: intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.replicas'}) + ' x',
-                             description: (
-                               <Space>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.requests.cpu'})}
-                                   value={data.brokerSpec?.requests?.cpu}/>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.requests.memory'})}
-                                   value={data.brokerSpec?.requests?.memory}/>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.limits.cpu'})}
-                                   value={data.brokerSpec?.limits?.cpu}/>
-                                 <Statistic
-                                   title={intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.limits.memory'})}
-                                   value={data.brokerSpec?.limits?.memory}/>
-                               </Space>
-                             ),
-                           }}
-                           footer={
-                             <div>{intl.formatMessage({id: 'pages.project.doris.template.steps.component.base.image'}) + ": " + data.brokerSpec?.image ? data.brokerSpec?.image : '-'}</div>
-                           }
-            />
-          </ProCard>
-        </ProCard.Group>
-        <Divider type={'horizontal'}/>
-        <ProCard.Group title={intl.formatMessage({id: 'pages.project.doris.instance.detail.yaml'})}
-                       direction={responsive ? 'column' : 'row'}>
-          <ProCard bordered>
-            <Editor
-              width="730"
-              height="600px"
-              language="yaml"
-              theme="vs-white"
-              value={yaml}
-              options={{
-                selectOnLineNumbers: true,
-                readOnly: true,
-                minimap: {
-                  enabled: false
-                }
-              }}
-              onMount={handleEditorDidMount}
-            />
-          </ProCard>
-        </ProCard.Group>
-      </RcResizeObserver>
+      <DorisInstanceDetailAction data={data}/>
+      <Divider type={'horizontal'}/>
+      <DorisInstanceDetailComponent data={data}/>
+      <Divider type={'horizontal'}/>
+      <DorisInstanceDetailAccess data={data}/>
+      <Divider type={'horizontal'}/>
+      <DorisInstanceDetailYAML data={data}/>
     </PageContainer>
   );
 }
